vpx_ios/VPXTest: Fix warnings.
1. Avoid loss of precision warnings by using obj-C integer types.
2. Because of 1: Use explicit casts in NSLog when logging NS* integer
   types. (Without the casts this introduces more warnings-- because
   NS* int type sizes change based on target)

+++ vpx_ios/VPXTest/ViewController.h
@@ -10,9 +10,9 @@
 
 @interface ViewController : UIViewController<UITableViewDataSource> {
  @private
-  int64_t downloadFileIndex_;
-  int64_t playFileIndex_;
-  int64_t selectedFileIndex_;
+  NSInteger downloadFileIndex_;
+  NSInteger playFileIndex_;
+  NSInteger selectedFileIndex_;
   NSArray *testFiles_;
   NSURL *testFileDownloadURL_;
   NSURL *cacheDirectoryPath_;
